+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Plan Downloads [#299]
 -
 ### Added
+- Added support for question type JSON objects [#535]
 - Added TinyMCE Editor component and added it to the styleguide [#462]
 - Added a Load more button to the Add new section page [#450]
 - PublishedSections graphQL query [#448]